update Apec jobboard
This patch updates the filter_contract_fr() function from utilities.py to
take in account the new job type description from the Apec website, now
describing a CDI as `2 en CDI' and a CDD as `1 en CDD de 3 mois'.
A minor modification is done to the report_generate() function so it detects
properly a CDD and flag the line as it.