Report js filtering: add cache to SalaryFilter.
Add a cache system (currently, a javascript array with string index,
which is in fact an hashmap) to the computation of the salary range of
an offer from its salary description.
Additionnally, update the table only when the salary handles are
released and not at every change. The salary filter feedback (reading
"De xk€ à yk€") is still updated at each change.
Additionnally, fix doc.